+++ tests/unit/components/paper-autocomplete-test.js
@@ -1,4 +1,5 @@
 import { moduleForComponent, test } from 'ember-qunit';
+import Ember from 'ember';
 
 moduleForComponent('paper-autocomplete', 'Unit | Component | paper autocomplete', {
   // Specify the other units that are required for this test
@@ -41,4 +42,57 @@ test('it propagates placeholder to input box', function(assert) {
   assert.equal(component.$().find('input').attr('placeholder'), 'Testing', 'Sets correct placeholder on input box.');
 });
 
+test('full-text search', function(assert) {
+  assert.expect(1);
+
+  // Creates the component instance
+  let component = this.subject();
+  // Renders the component to the page
+  this.render();
+  Ember.run(()=>{
+    component.set('source', [
+      {
+        name:'full-text search'
+      },
+      {
+        name:'not full-text search'
+      },
+      {
+        name:'search'
+      }
+    ]);
+    component.set('fullTextSearch',true);
+    component.set('lookupKey','name');
+    component.set('searchText','search');
+    component.set('debouncedSearchText','search');
+  });
+  assert.equal(component.get('suggestions').length,3);
+});
+
+test('non full-text search', function(assert) {
+  assert.expect(1);
+
+  // Creates the component instance
+  let component = this.subject();
+  // Renders the component to the page
+  this.render();
+  Ember.run(()=>{
+    component.set('source', [
+      {
+        name:'full-text search'
+      },
+      {
+        name:'not full-text search'
+      },
+      {
+        name:'search'
+      }
+    ]);
+    component.set('fullTextSearch',false);
+    component.set('lookupKey','name');
+    component.set('searchText','search');
+    component.set('debouncedSearchText','search');
+  });
+  assert.equal(component.get('suggestions').length,1);
+});