+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Ember Paper Changelog
 
-### master
+### 0.2.12
 
 - [#307](https://github.com/miguelcobain/ember-paper/pull/307) Add paper-card title components
 - [#283](https://github.com/miguelcobain/ember-paper/pull/283) Adds support for `fullTextSearch` attribute on `{{paper-autocomplete}}`. Enables passing Promises to the `model` attribute on `{{paper-autocomplete}}`. Docs updated.